feat: Install run-code plugin
 A simple plugin to quickly run a snippet of code without ever leaving neovim

Usage
=====

This is still a work in progress. So, PRs are welcome and appreciated. As of now, this plugin can do the following things

| Command | Description |
|---------|-------------|
| `:RunCodeSelected` | Run code that is visually selected |
| `:RunCodeFile` | Run the entire file |
| `:RunCodeBlock` | Run a code snippet present in markdown (place the cursor within the snippet and run the command) |
| `:RunCodeSetCustomCmd <alias>` | Set a custom commands to use with `:RunCodeCustomCmd`. This custom command would be used instead of the default run command used for a specific language |
| `:RunCodeCustomCmd <alias>` | Run the custom command stored at the given alias |
| `:RunCodeLauncher` | Launches a [telescope](https://github.com/nvim-telescope/telescope.nvim) interface to manage custom commands. Read the docs [here](https://github.com/arjunmahishi/run-code.nvim/wiki/Run-code-launcher) |

Custom commands
===============

Custom commands are snippets of code you can map to an alias to access and run
them easily. Custom commands can be created using either `:RunCodeSetCustomCmd`
or `RunCodeLauncher`. `:RunCodeLauncher` uses a nice telescope interface which
makes it easy to manage the commands. So, `:RunCodeLauncher` is recommended
over `:RunCodeSetCustomCmd`.

Running `:RunCodeSetCustomCmd <alias>` launches a special buffer which allows
you to create a custom command snippet. This buffer automatically closes on
saving. This remains same even if you are using `:RunCodeLauncher`. Once the
command snippet is saved, it is available to be executed using the alias you
have mapped it to.